+++ lib/discourse_topic_voting/categories_controller_extension.rb
@@ -5,7 +5,9 @@ module DiscourseTopicVoting
     def category_params
       @vote_enabled ||=
         !!ActiveRecord::Type::Boolean.new.cast(params[:custom_fields]&.delete(:enable_topic_voting))
+
       category_params = super
+
       if @vote_enabled && !@category&.category_setting
         category_params[:category_setting_attributes] = {}
       elsif !@vote_enabled && @category&.category_setting
@@ -14,6 +16,7 @@ module DiscourseTopicVoting
           _destroy: "1",
         }
       end
+
       category_params
     end
   end
